+++ CHANGELOG.md
@@ -7,3 +7,4 @@ Date: 2016-05-03
 - Better ScreenScraper cleaning
 - Some messages fixed
 - Bugfix: ScreenScraper genres escaped (fixes Run & gun)
+- Apply score even on match by filename/md5 (prevents bad rom association on ScreenScraper)